Restructure test examples
Remove generated files, add src-gen to projects without XtxtUML files,
simplify directory nesting in `MoonLander`, reformat and fix errors in
`validation`.
