+++ examples/tests/MoonLander/src/hu/eltesoft/moonlander/model/ControlCycleSignal.java
