+++ examples/tests/MoonLander/src/hu/eltesoft/moonlander/model/ControlSignal.java
