+++ examples/tests/airport/readme.txt
@@ -1,13 +1,3 @@
-Dear User!
-
-This project (demo/airport) is a still work in progress demo project 
-for txtUML tool. We wanted to share it because it uses a lot of 
-features of the tool and also uses the model to run a GUI with 
-a Glue layer.
-Other then this is a big project and a good proof of the tool 
-working it still needs to evolve as not every feature is 
-presented on the GUI. Therefore there may be "TODO" lines 
-scattered in the project. We will keep working on this demo.
-
-Regards,
-  txtUML team
+This example is under construction. It was made available as its multilayered
+architecture — where the model defines the business logic of the application —
+provides a more advanced view on modeling with txtUML.
