+++ examples/tests/hu.elte.txtuml.examples.printer/.gitignore
