+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/AssociationWrongInnerClass.java
@@ -6,7 +6,7 @@ public class AssociationWrongInnerClass extends Association {
 
 	class X {
 	}
-	
+
 	class Y {
 	}
