+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/CompositionNotExactlyOneContainer.java
@@ -8,7 +8,7 @@ class CompositionNoContainer extends Composition {
 
 	public class a extends End<Any<A>> {
 	}
-	
+
 	public class b extends End<Any<A>> {
 	}
 }
@@ -17,18 +17,18 @@ class CompositionBothContainer extends Composition {
 
 	public class a extends ContainerEnd<A> {
 	}
-	
+
 	public class b extends ContainerEnd<A> {
 	}
-	
+
 }
 
 class CompositionOK extends Composition {
 
 	public class a extends ContainerEnd<A> {
 	}
-	
+
 	public class b extends End<Any<A>> {
 	}
-	
+
 }
