+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/DataTypeFieldNotFinal.java
@@ -5,7 +5,7 @@ import hu.elte.txtuml.api.model.DataType;
 public class DataTypeFieldNotFinal extends DataType {
 
 	int a;
-	
+
 	final String b = "";
 
 }
