+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/DataTypeInvalidFieldType.java
@@ -7,13 +7,13 @@ import hu.elte.txtuml.examples.validation.model.helpers.MyDataType;
 public class DataTypeInvalidFieldType extends DataType {
 
 	final int a = 0;
-	
+
 	final String b = "";
-	
+
 	final MyDataType c = new MyDataType();
-	
+
 	final Object x = null;
-	
+
 	final A y = null;
 
 }
