+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/DataTypeInvalidMethod.java
@@ -5,13 +5,13 @@ import hu.elte.txtuml.api.model.DataType;
 public class DataTypeInvalidMethod extends DataType {
 
 	final int a;
-	
+
 	public DataTypeInvalidMethod() {
 		a = 43;
 	}
-	
+
 	public void f() {
-		
+
 	}
 
 }
