+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidClass.java
@@ -1,4 +1,4 @@
 package hu.elte.txtuml.examples.validation.model;
-public class InvalidClass {
 
+public class InvalidClass {
 }
