+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidFieldType.java
@@ -1,4 +1,5 @@
 package hu.elte.txtuml.examples.validation.model;
+
 import hu.elte.txtuml.api.model.ModelClass;
 import hu.elte.txtuml.examples.validation.model.helpers.A;
 import hu.elte.txtuml.examples.validation.model.helpers.ExternalClass;
