+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidModifier.java
@@ -1,13 +1,15 @@
 package hu.elte.txtuml.examples.validation.model;
+
 import hu.elte.txtuml.api.model.ModelClass;
-import hu.elte.txtuml.api.model.Signal;
 
 public class InvalidModifier extends ModelClass {
-	
-	synchronized void f() {}
-	
-	static final void g() {}
-	
+
+	synchronized void f() {
+	}
+
+	static final void g() {
+	}
+
 	static int i;
-	
+
 }
