+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidTypeInModelClass.java
@@ -1,8 +1,9 @@
 package hu.elte.txtuml.examples.validation.model;
+
 import hu.elte.txtuml.api.model.ModelClass;
 
 public class InvalidTypeInModelClass extends ModelClass {
-	
+
 	Object o;
 
 }
