+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidTypeInSignal.java
@@ -1,4 +1,5 @@
 package hu.elte.txtuml.examples.validation.model;
+
 import hu.elte.txtuml.api.model.Signal;
 import hu.elte.txtuml.examples.validation.model.helpers.A;
