+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/MethodInSignal.java
@@ -1,11 +1,12 @@
 package hu.elte.txtuml.examples.validation.model;
+
 import hu.elte.txtuml.api.model.Signal;
 
 public class MethodInSignal extends Signal {
 
 	public MethodInSignal() {
 	}
-	
+
 	public void f() {
 	}
