+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/StateMethodNotCorrect.java
@@ -14,12 +14,12 @@ public class StateMethodNotCorrect extends ModelClass {
 	public class St extends State {
 		public void entry() {
 		}
-		
+
 		public void exit(int x, int y) {
 		}
-		
+
 		public void foo(int x) {
-			
+
 		}
 	}
