+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/helpers/ExternalClass.java
@@ -5,5 +5,4 @@ import hu.elte.txtuml.api.model.ModelClass;
 
 @External
 public class ExternalClass extends ModelClass {
-
 }
