+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/helpers/MyDataType.java
@@ -3,5 +3,4 @@ package hu.elte.txtuml.examples.validation.model.helpers;
 import hu.elte.txtuml.api.model.DataType;
 
 public class MyDataType extends DataType {
-
 }
