+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/helpers/Sig.java
@@ -3,5 +3,4 @@ package hu.elte.txtuml.examples.validation.model.helpers;
 import hu.elte.txtuml.api.model.Signal;
 
 public class Sig extends Signal {
-
 }
