+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/sequencediagram/BaseSequence.java
@@ -1,12 +1,11 @@
 package hu.elte.txtuml.examples.validation.sequencediagram;
 
 import hu.elte.txtuml.api.model.Action;
+import hu.elte.txtuml.api.model.seqdiag.Position;
+import hu.elte.txtuml.api.model.seqdiag.SequenceDiagram;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.A;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.AToB;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.B;
-import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.TestSig;
-import hu.elte.txtuml.api.model.seqdiag.Position;
-import hu.elte.txtuml.api.model.seqdiag.SequenceDiagram;
 
 public abstract class BaseSequence extends SequenceDiagram {
 
@@ -17,7 +16,7 @@ public abstract class BaseSequence extends SequenceDiagram {
 
 	@Override
 	public void initialize() {
-		lifeline1 = Action.create(A);
+		lifeline1 = Action.create(A.class);
 		lifeline2 = new B();
 		Action.link(AToB.ASide.class, lifeline1, AToB.BSide.class, lifeline2);
 	}
