+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/sequencediagram/EmptyRun.java
@@ -3,7 +3,7 @@ package hu.elte.txtuml.examples.validation.sequencediagram;
 import hu.elte.txtuml.api.model.seqdiag.SequenceDiagram;
 
 public class EmptyRun extends SequenceDiagram {
-	
+
 	@Override
 	public void initialize() {
 	}
