+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/sequencediagram/InvalidActionCall.java
@@ -1,6 +1,7 @@
 package hu.elte.txtuml.examples.validation.sequencediagram;
 
 import hu.elte.txtuml.api.model.Action;
+import hu.elte.txtuml.api.model.seqdiag.Sequence;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.A;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.TestSig;
