+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/sequencediagram/ValidDiagram.java
@@ -1,9 +1,9 @@
 package hu.elte.txtuml.examples.validation.sequencediagram;
 
 import hu.elte.txtuml.api.model.Action;
-import hu.elte.txtuml.api.model.ModelClass;
 import hu.elte.txtuml.api.model.seqdiag.Position;
 import hu.elte.txtuml.api.model.seqdiag.Sequence;
+import hu.elte.txtuml.api.model.seqdiag.SequenceDiagram;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.A;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.AToB;
 import hu.elte.txtuml.examples.validation.sequencediagram.testmodel.B;
@@ -87,31 +87,6 @@ class Container extends SequenceDiagram {
 	@Override
 	public void run() {
 		diagr1.run();
-	}
-
-}
-
-public class ValidDiagramWithMethodInvocations extends SequenceDiagram {
-
-	@Position(10)
-	private A lifeline1;
-
-	public B lifeline2;
-
-	@Override
-	public void initialize() {
-		lifeline1 = new A();
-		lifeline2 = new B();
-		Action.link(AToB.ASide.class, lifeline1, AToB.BSide.class, lifeline2);
-	}
-
-	@Override
-	public void run() {
-		someMethod();
-	}
-
-	private void someMethod() {
-		Sequence.fromActor(new TestSig(), lifeline2);
 	}
 
 }
