+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/sequencediagram/testmodel/C.java
@@ -1,0 +1,6 @@
+package hu.elte.txtuml.examples.validation.sequencediagram.testmodel;
+
+import hu.elte.txtuml.api.model.ModelClass;
+
+public class C extends ModelClass {
+}
