--- examples/tests/hu.elte.txtuml.export.uml2.tests.models/.gitignore
@@ -1,2 +1,0 @@
-!src-gen
-*.java._trace
