--- examples/tests/hu.elte.txtuml.export.uml2.tests.models/src-gen/hu/elte/txtuml/export/uml2/tests/models/mixed/B.java
@@ -1,7 +1,0 @@
-package hu.elte.txtuml.export.uml2.tests.models.mixed;
-
-import hu.elte.txtuml.api.model.ModelClass;
-
-@SuppressWarnings("all")
-public class B extends ModelClass {
-}
