--- examples/tests/hu.elte.txtuml.export.uml2.tests.models/src-gen/hu/elte/txtuml/export/uml2/tests/models/mixed/package-info.java
@@ -1,4 +1,0 @@
-@Model("mixed")
-package hu.elte.txtuml.export.uml2.tests.models.mixed;
-
-import hu.elte.txtuml.api.model.Model;
