+++ examples/tests/hu.elte.txtuml.seqdiag.tests.models/.gitignore
@@ -1,1 +1,1 @@
-!src-gen
+!src-gen/.gitkeep
