+++ examples/tests/hu.elte.txtuml.seqdiag.tests.models/src/hu/elte/txtuml/seqdiag/tests/models/sequences/ParFragment.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.export.plantuml.tests.models.sequences;
+package hu.elte.txtuml.seqdiag.tests.models.sequences;
 
 import hu.elte.txtuml.api.model.seqdiag.Sequence;
 import hu.elte.txtuml.seqdiag.tests.models.testmodel.TestSig;
